+++ src/main/kotlin/uk/co/ben_gibson/git/link/git/Host.kt
@@ -1,5 +1,7 @@
 package uk.co.ben_gibson.git.link.git
 
+import uk.co.ben_gibson.git.link.GitLinkBundle
+import uk.co.ben_gibson.git.link.ui.Icons
 import uk.co.ben_gibson.git.link.url.template.UrlTemplates
 import java.net.URL
 import java.util.UUID
@@ -12,13 +14,21 @@ val HOST_ID_BITBUCKET_SERVER: UUID = UUID.fromString("dba5941d-821c-49b3-83b0-75
 val HOST_ID_GOGS: UUID = UUID.fromString("fd2d9cfc-1eef-4b1b-80bd-b02def58576c")
 val HOST_ID_GITEA: UUID = UUID.fromString("e0f86390-1091-4871-8aeb-f534fbc99cf0")
 val HOST_ID_AZURE: UUID = UUID.fromString("83008277-73fa-4faa-b9b2-0a60fecb030e")
+val HOST_ID_CHROMIUM: UUID = UUID.fromString("97bf87bc-99ef-4e1f-8d37-7948a2082df4")
 
 sealed class Host(val id: UUID, val displayName: String, val icon: Icon, val baseUrl: URL? = null)
 
 class TemplatedHost(
-        id: UUID,
-        displayName: String,
-        icon: Icon,
-        val urlTemplate: UrlTemplates,
-        baseUrl: URL? = null
+    id: UUID,
+    displayName: String,
+    icon: Icon,
+    val urlTemplate: UrlTemplates,
+    baseUrl: URL? = null
 ) : Host(id, displayName, icon, baseUrl)
+
+class ChromiumHost() : Host(
+    UUID.fromString("97bf87bc-99ef-4e1f-8d37-7948a2082df4"),
+    GitLinkBundle.message("hosts.chromium.name"),
+    Icons.CHROMIUM,
+    URL("https://chromium.googlesource.com")
+)