+++ src/main/kotlin/uk/co/ben_gibson/git/link/pipeline/ForceHttpsMiddleware.kt
@@ -2,20 +2,18 @@ package uk.co.ben_gibson.git.link.pipeline
 
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
-import com.intellij.openapi.project.Project
 import uk.co.ben_gibson.git.link.settings.ProjectSettings
 import uk.co.ben_gibson.git.link.url.toHttps
 import java.net.URL
-import uk.co.ben_gibson.git.link.Context
 
 @Service
 class ForceHttpsMiddleware : Middleware {
-    override val priority = 3
+    override val priority = 30
 
-    override fun invoke(project: Project, context: Context, next: () -> URL?) : URL? {
+    override fun invoke(pass: Pass, next: () -> URL?) : URL? {
         val url = next() ?: return null
 
-        val settings = project.service<ProjectSettings>()
+        val settings = pass.project.service<ProjectSettings>()
 
         if (settings.forceHttps) {
             return url.toHttps()